third_party/blink/web_tests/http/tests/security/mixedContent/insecure-async-post-xhr-blocked-expected.txt

CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-async-xhr-post.html' was loaded over HTTPS, but requested an insecure XMLHttpRequest endpoint 'http://example.test:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=*'. This request has been blocked; the content must be served over HTTPS.
This is a testharness.js-based test.
PASS Opens an HTTPS window that loads insecure data via XHR. We should trigger a mixed content callback and block the load.
Harness: the test ran to completion.
